updated rendering scheme (see details)
Renderable, Logicable are defined by the application itself now, not by
the Engine. instead, the Engine provides the Runner trait, which can be
implemented by the application for better customization of rendering and
updating.

vertex buffer descriptors and rendering pipelines are handled by the
modules that correlate to different pipelines. for example, sprites and
3D blocks have their own pipelines, and their modules will take care of
and export functions that return vertex buffer descriptors and said
pipelines